+++ src/main/java/com/robertx22/age_of_exile/aoe_data/GeneratedData.java
@@ -27,6 +27,7 @@ import com.robertx22.age_of_exile.aoe_data.database.stats.AutoDatapackStats;
 import com.robertx22.age_of_exile.aoe_data.database.stats.DatapackStatAdder;
 import com.robertx22.age_of_exile.aoe_data.database.stats.SpellDependentDatapackStatAdder;
 import com.robertx22.age_of_exile.aoe_data.database.unique_gears.UniqueGearReg;
+import com.robertx22.age_of_exile.database.data.game_balance_config.GameBalanceConfig;
 import com.robertx22.age_of_exile.database.registrators.Tiers;
 import com.robertx22.age_of_exile.mmorpg.MMORPG;
 
@@ -79,6 +80,8 @@ public class GeneratedData {
 
             new SkillGemsAdder().registerAll();
             new RandomSkillGemStatsAdder().registerAll();
+
+            new GameBalanceConfig().addToSerializables();
 
         }
     }