+++ src/main/java/com/robertx22/age_of_exile/config/forge/ServerContainer.java
@@ -25,14 +25,9 @@ public class ServerContainer {
 
     public double EXTRA_MOB_STATS_PER_LEVEL = 0.02F;
 
-    public double TALENT_POINTS_AT_MAX_LEVEL = 125;
-    public double STARTING_TALENT_POINTS = 1;
-
     public double VANILLA_MOB_DMG_AS_EXILE_DMG = 0.5F;
     public double PVP_DMG_MULTI = 1F;
 
     public List<String> IGNORED_ENTITIES = Arrays.asList("minecraft:bat");
-
-    public int MAX_LEVEL = 100;
 
 }