+++ src/main/java/com/robertx22/age_of_exile/database/data/BaseRuneGem.java
@@ -1,13 +1,12 @@
 package com.robertx22.age_of_exile.database.data;
 
+import com.robertx22.age_of_exile.database.data.game_balance_config.GameBalanceConfig;
 import com.robertx22.age_of_exile.database.data.gear_types.bases.BaseGearType.SlotFamily;
 import com.robertx22.age_of_exile.uncommon.interfaces.IWeighted;
 import net.minecraft.util.math.MathHelper;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import static com.robertx22.age_of_exile.config.forge.ModConfig.get;
 
 public class BaseRuneGem implements IGUID, IWeighted {
 
@@ -28,11 +27,11 @@ public class BaseRuneGem implements IGUID, IWeighted {
     public float effective_level;
 
     public int getEffectiveLevel() {
-        return (int) (MathHelper.clamp(get().Server.MAX_LEVEL * effective_level, 1, get().Server.MAX_LEVEL));
+        return (int) (MathHelper.clamp(GameBalanceConfig.get().MAX_LEVEL * effective_level, 1, GameBalanceConfig.get().MAX_LEVEL));
     }
 
     public int getReqLevel() {
-        return (int) (MathHelper.clamp(get().Server.MAX_LEVEL * required_item_level, 1, get().Server.MAX_LEVEL));
+        return (int) (MathHelper.clamp(GameBalanceConfig.get().MAX_LEVEL * required_item_level, 1, GameBalanceConfig.get().MAX_LEVEL));
     }
 
     public final List<StatModifier> getFor(SlotFamily sfor) {