+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/StatScaling.java
@@ -1,6 +1,6 @@
 package com.robertx22.age_of_exile.database.data.stats;
 
-import com.robertx22.age_of_exile.config.forge.ModConfig;
+import com.robertx22.age_of_exile.database.data.game_balance_config.GameBalanceConfig;
 
 public enum StatScaling {
 
@@ -13,7 +13,13 @@ public enum StatScaling {
     NORMAL {
         @Override
         public float scale(float val, int lvl) {
-            return val * ModConfig.get().statScalings.NORMAL_STAT_SCALING.getMultiFor(lvl);
+            return val * GameBalanceConfig.get().NORMAL_STAT_SCALING.getMultiFor(lvl);
+        }
+    },
+    STAT_REQ {
+        @Override
+        public float scale(float val, int lvl) {
+            return val * GameBalanceConfig.get().STAT_REQ_SCALING.getMultiFor(lvl);
         }
     },
     LINEAR {
@@ -25,7 +31,7 @@ public enum StatScaling {
     SLOW {
         @Override
         public float scale(float val, int lvl) {
-            return val * ModConfig.get().statScalings.SLOW_STAT_SCALING.getMultiFor(lvl);
+            return val * GameBalanceConfig.get().SLOW_STAT_SCALING.getMultiFor(lvl);
         }
     };