+++ src/main/java/com/robertx22/age_of_exile/loot/blueprints/bases/LevelPart.java
@@ -1,6 +1,6 @@
 package com.robertx22.age_of_exile.loot.blueprints.bases;
 
-import com.robertx22.age_of_exile.config.forge.ModConfig;
+import com.robertx22.age_of_exile.database.data.game_balance_config.GameBalanceConfig;
 import com.robertx22.age_of_exile.loot.blueprints.GearBlueprint;
 import com.robertx22.age_of_exile.loot.blueprints.ItemBlueprint;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.RandomUtils;
@@ -45,6 +45,7 @@ public class LevelPart extends BlueprintPart<Integer, ItemBlueprint> {
 
         finalLvl = MathHelper.clamp(finalLvl, minLevel, maxLevel);
 
-        return MathHelper.clamp(finalLvl, this.minLevel, ModConfig.get().Server.MAX_LEVEL);
+        return MathHelper.clamp(finalLvl, this.minLevel, GameBalanceConfig.get().MAX_LEVEL);
     }
+
 }